Remove tests for relative/absolute paths, no longer relevant. Move normalize_path to _legacy.
