+++ importlib_resources/_common.py
@@ -6,7 +6,7 @@ import contextlib
 import types
 import importlib
 
-from typing import Union, Any, Optional
+from typing import Union, Optional
 from .abc import ResourceReader, Traversable
 
 from ._compat import wrap_spec
@@ -21,19 +21,6 @@ def files(package):
     Get a Traversable resource from a package
     """
     return from_package(get_package(package))
-
-
-def normalize_path(path):
-    # type: (Any) -> str
-    """Normalize a path by ensuring it is a string.
-
-    If the resulting string contains path separators, an exception is raised.
-    """
-    str_path = str(path)
-    parent, file_name = os.path.split(str_path)
-    if parent:
-        raise ValueError(f'{path!r} must be only a file name')
-    return file_name
 
 
 def get_resource_reader(package):
